+++ src/test/lib/examples/path-norm/node-protocol/always/result.ts
@@ -1,0 +1,3 @@
+import { A } from 'aaa';
+import { C } from 'node:aaa';
+import { B, D } from 'node:assert';