+++ src/test/lib/examples/path-norm/node-protocol/preserve/origin.ts
@@ -1,0 +1,4 @@
+import {A} from 'aaa'
+import {B} from 'assert'
+import {C} from 'node:aaa'
+import {D} from 'node:assert'